requirements.txt
annotated-types
anyio
blinker
CacheControl
cachetools
certifi
cffi
charset-normalizer
click
cryptography
distro
emoji
firebase-admin
Flask
flask-cors
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-cloud-core
google-cloud-firestore
google-cloud-storage
google-crc32c
google-resumable-media
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
idna
itsdangerous
Jinja2
jiter
MarkupSafe
msgpack
numpy
openai
opencv-python
pillow
pip
praw
prawcore
proto-plus
protobuf
pyasn1
pyasn1_modules
pycparser
pydantic
pydantic_core
pydeps
PyJWT
pyparsing
python-dotenv
requests
rsa
sniffio
stdlib-list
tqdm
typing_extensions
typing-inspection
update-checker
uritemplate
urllib3
websocket-client
Werkzeug
wheel
attrs
pytesseract
algoliasearch
google-cloud-vision